requirements.txt
Django==3.2
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-environ==0.4.5
django-filter==2.4.0
django-rest-knox==4.1.0
django-storages==1.11.1
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.1.2
psycopg2==2.8.6
django-nested-inline==0.4.2
PyVimeo==1.1.0
python-dateutil==2.8.1
xxhash==2.0.2
django-redis==4.12.1
boto3==1.17.69